Monitor sensor data on TrainingFragment
The sensor monitor on TrainingFragment is implemented, it shows the newest value from all registered sensor every 100 CountDownTimer ticks.

To make this possible, the implementation of LogSensorService is also changed. The time based while loop (which is ugly) is changed to onSensorDataReady callback method, invoked when SensorReader finishes reading all registered sensor  values into buffer.

Note that the implementation of readyToRead() on SensorReader (which calls the onSensorDataReady if it returns true) waits for every buffer is filled. This could cause a problem if the sensors sampling rate is significantly different.