Make sure SensorData is not corrupted before SensorDataSequence commit
In some rare cases, the sensors are not available so the data is corrupted with nulls. To prevent bizzare things from happening, corruped sensor data is ignored.

This corruption prevention is done before committing SensorData to SensorDataSequence. Note that current implementation is quite expensive, so there might be better approach to prevent corruption.