Make LogSensorService as a foreground service
A notification is created (as required by foreground service) containing progress bar and time lelft text. When pressed, this notification will resume TrainingActivity. To make that happen, The launchMode of TrainingActivity is set to singleInstance. Otherwise it will create a new TrainingActivity.

The whole messy things is that I need to change the method to pass activity information. So the HumanActivity class is created as a vessel. Now all information of activity (like name and icon) should be passed on HumanActivity object.