Change the design of HumanActivity and HumanActivityHistory
The HumanActivityRegister is deleted. I think that using a whole class as a list of activity type is silly. So instead, I create a switch case in HumanActivity to determine the resource for each activity.

I also move the ActivityId from HumanActivityRegister to Id in HumanActivity. I also improve it with an internal map of int-to-Id, complete with a static method to get Id from int.