Create PredictionService
PredictionService is a service to make prediction from sensor data. It is **NOT** ready to use yet. Some method is not yet implemented. This service also calls a native method to classify activity with TensorFlow (also not yet implemented).

To avoid diplication,  SensorService is created as a base of LogSensorService and Prediction service., which both need a similar sensor reading routine.