Add metadata in the first line of log file
A metadata is added to ease the file parsing process. It includes:
- Log type
- Number of sensor used
- Total number of axis from all sensor
- Number of sensor data entry

With this approach, all sensor data is written at the end of LogSensorService process, so all sensor data is stored in memory while the service is running. I don't know if its a good idea, I'm afraid it consumes too much memory for long running process.