Ask device placement before training
Add a spinner in TrainingChooserDialogFragment to choose device placement for training process. This information is required for analyzing the performance of recognition system.

This information is passed by an intent extra from TraiingChooserDialogFragment to TrainingActivity, then from TrainingActivity to LogSensorService. The LogSensorService will write that information in the log file.