Write log to database after training
This database is needed for storing information of the log's status. Right after training , status will set to PENDING, then the log will be sent to server and the status will be changed to SENT (not implemented yet) . This way, the app can figure out which log to send to server.